+++ lib/src/sp3d_world.dart
@@ -13,8 +13,8 @@ import 'sp3d_paint_image.dart';
 /// First edition creation date 2021-09-30 14:58:34
 ///
 class Sp3dWorld {
-  final String className = 'Sp3dWorld';
-  final String version = '6';
+  String get className => 'Sp3dWorld';
+  String get version => '6';
   List<Sp3dObj> objs;
 
   // 以下はディープコピーなどが不要な一時変数。